+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Added
 - `EachRule`
+- Made `Validator` a callable class
+
+### Deprecated
+- `Validator.build()`
 
 ### Changed
 - Moved all files to src folder. Use `import 'package:flrx_validator/validator.dart';` *BREAKING CHANGE*